ai-coding-course-claude/lessons/lesson-03-perfect-prompt/src/claude-3/pyproject.toml
[project]
name = "transcript-analyzer"
version = "0.1.0"
description = "Python-based transcript analysis tool using Google Gemini AI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "matplotlib>=3.7.0",
]

[project.scripts]
main = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]